@@ -1,16 +1,15 @@
 name: heap_analytics
 description: Heap Analytics
 version: 0.1.2
-author: Kurt Inge Smådal <kurtinge@gmail.com>
-homepage:
+homepage: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.0
+  http: ^0.13.3
 
 dev_dependencies:
   flutter_test:
